Move app_info.py into RimSort.utils
