+++ RimSort/utils/app_info.py
