+++ view/main_window.py
@@ -10,7 +10,7 @@ from watchdog.observers.api import BaseObserver
 from RimSort.controllers.menu_bar_controller import MenuBarController
 from RimSort.controllers.settings_controller import SettingsController
 from RimSort.models.settings import Settings
-from util.app_info import AppInfo
+from RimSort.utils.app_info import AppInfo
 from util.system_info import SystemInfo
 from util.watchdog import RSFileSystemEventHandler
 from view.game_configuration_panel import GameConfiguration
